+++ tests/nn_primitives/test_nnp_convolution.nim
@@ -37,10 +37,10 @@ proc main() =
       check: input.conv2d(kernel, bias, padding=(1,1)) == target
 
       let
-        finput = input.astype(float32)
-        fkernel = kernel.astype(float32)
-        fbias = bias.astype(float32)
-        ftarget = target.astype(float32)
+        finput = input.asType(float32)
+        fkernel = kernel.asType(float32)
+        fbias = bias.asType(float32)
+        ftarget = target.asType(float32)
 
       test "Simple Conv2D [Im2ColGEMM]":
         check: finput.conv2d(fkernel, fbias, padding=(1,1)).mean_absolute_error(ftarget) <= 1e-7'f32
@@ -126,10 +126,10 @@ proc main() =
       check: input.conv2d(kernel, bias, padding=(1,1), stride=(2,2)) == target
 
       let
-        finput = input.astype(float32)
-        fkernel = kernel.astype(float32)
-        fbias = bias.astype(float32)
-        ftarget = target.astype(float32)
+        finput = input.asType(float32)
+        fkernel = kernel.asType(float32)
+        fbias = bias.asType(float32)
+        ftarget = target.asType(float32)
 
       check: finput.conv2d(fkernel, fbias, padding=(1,1), stride=(2,2)) == ftarget
 
@@ -145,9 +145,9 @@ proc main() =
       let output = conv2d(input, kernel, bias, padding, stride)
 
       let
-        dinput = input.astype(float)
-        dkernel = kernel.astype(float)
-        dbias = bias.astype(float)
+        dinput = input.asType(float)
+        dkernel = kernel.asType(float)
+        dbias = bias.asType(float)
 
       let
         target_grad_input = dinput.numerical_gradient(
@@ -163,18 +163,18 @@ proc main() =
       test "Conv2D Forward + Backward [Im2ColGEMM]":
         conv2d_backward(input, kernel, bias, padding, stride,
                         grad_output, grad_input, grad_weight, grad_bias)
-        check: grad_bias.astype(float).mean_relative_error(target_grad_bias) < 1e-6
-        check: grad_weight.astype(float).mean_relative_error(target_grad_weight) < 1e-6
-        check: grad_input.astype(float).mean_relative_error(target_grad_input) < 1e-6
+        check: grad_bias.asType(float).mean_relative_error(target_grad_bias) < 1e-6
+        check: grad_weight.asType(float).mean_relative_error(target_grad_weight) < 1e-6
+        check: grad_input.asType(float).mean_relative_error(target_grad_input) < 1e-6
 
       when defined(nnpack):
         test "Conv2D Forward + Backward [NNPack]":
           conv2d_backward(input, kernel, bias, padding, stride,
                           grad_output, grad_input, grad_weight, grad_bias,
                           algorithm=Conv2DAlgorithm.NNPackAuto)
-          check: grad_bias.astype(float).mean_relative_error(target_grad_bias) < 1e-6
-          check: grad_weight.astype(float).mean_relative_error(target_grad_weight) < 1e-6
-          check: grad_input.astype(float).mean_relative_error(target_grad_input) < 1e-6
+          check: grad_bias.asType(float).mean_relative_error(target_grad_bias) < 1e-6
+          check: grad_weight.asType(float).mean_relative_error(target_grad_weight) < 1e-6
+          check: grad_input.asType(float).mean_relative_error(target_grad_input) < 1e-6
 
 
 main()
